+++ packages/template-retail-react-app/app/components/list-menu/index.js
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2024, Salesforce, Inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: BSD-3-Clause
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+export {ListMenu} from './list-menu'
+export {ListMenuItem} from './list-menu-item'
